GSTProject/ecommerceGST/requirements.txt
asgiref==3.2.10
defusedxml==0.6.0
diff-match-patch==20181111
Django==3.0.7
django-cors-headers==3.4.0
django-excel==0.0.10
django-import-export==2.2.0
djangorestframework==3.11.0
et-xmlfile==1.0.1
jdcal==1.4.1
lml==0.0.9
MarkupPy==1.14
numpy==1.19.0
odfpy==1.4.1
openpyxl==3.0.4
pandas==1.0.5
Pillow==7.2.0
psycopg2==2.8.5
pyexcel==0.6.2
pyexcel-io==0.5.20
pyexcel-webio==0.1.4
python-dateutil==2.8.1
pytz==2020.1
PyYAML==5.3.1
six==1.15.0
sqlparse==0.3.1
tablib==2.0.0
texttable==1.6.2
xlrd==1.2.0
xlutils==2.0.0
xlwt==1.3.0
